+++ src/dg_commons_dev/controllers/mpc/mpc_base_classes/full_mpc_base.py
@@ -1,7 +1,7 @@
 from typing import Tuple
 from abc import abstractmethod
-from dg_commons_dev.controllers.mpc.lateral_mpc_base import vehicle_params, LatMPCKinBase, LatMPCKinBaseParam
-from dg_commons_dev.controllers.cost_functions import *
+from dg_commons_dev.controllers.mpc.mpc_base_classes.lateral_mpc_base import vehicle_params, LatMPCKinBase, LatMPCKinBaseParam
+from dg_commons_dev.controllers.utils.cost_functions import *
 from dg_commons_dev.controllers.controller_types import LatAndLonController, LatAndLonControllerParam
