+++ src/dg_commons_dev/controllers/utils/cost_functions.py
