+++ src/dg_commons_dev/controllers/utils/discretization_techniques.py
